refactor: Change the folder of the project
