+++ build.gradle
@@ -5,7 +5,7 @@ plugins {
 }
 
 version = "0.1"
-group = "io.sc.checkout"
+group = "io.shoppingcard"
 
 repositories {
     mavenCentral()
@@ -16,7 +16,7 @@ micronaut {
     testRuntime("junit5")
     processing {
         incremental(true)
-        annotations("io.sc.checkout.*")
+        annotations("io.shoppingcard.*")
     }
 }
 
@@ -59,7 +59,7 @@ protobuf {
 }
 
 application {
-    mainClass.set("io.sc.checkout.Application")
+    mainClass.set("io.shoppingcard.Application")
 }
 
 java {
